requirements.txt
# Core dependencies for intersectional fairness research
pandas>=1.3.0
numpy>=1.20.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
scipy>=1.7.0
statsmodels>=0.12.0
PyYAML>=5.4.0

# Plotting and visualization
tueplots>=0.0.5

# Jupyter notebook support
jupyter>=1.0.0
ipython>=7.0.0

# Development dependencies (optional)
pytest>=6.0
black>=21.0
flake8>=3.8
